Redo build system
Redoes the build system so that asset files are built to .build, and
an asset map is built from their hashes if they have changed. Then
the build.ninja is rewritten with the versions, and a rule to copy
the versioned file to static. This way ninja can keep track of
everything.