Add a basic datatables style table to high_scores
This is a big change which adds datatables style functionality to the
high scores table. This will mean that the high scores can grow
indefinitely without killing the server every time it is requested.

There is only minimal functionality, further will be added later.

This also adds a basic bundle/require javascript system, and building of
javascript files.